rag_support_assistant/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

chromadb>=0.5.0
sentence-transformers>=2.7.0
pymupdf>=1.24.0

langgraph>=0.0.50

openai>=1.30.0